+++ src/index.js
@@ -6,7 +6,7 @@ import ErrorResponse from './protocol/control_layer/error_response/ErrorResponse
 import './protocol/control_layer/error_response/ErrorResponseSerializerV1'
 import './protocol/control_layer/error_response/ErrorResponseSerializerV2'
 import InstructionMessage from './protocol/tracker_layer/instruction_message/InstructionMessage'
-import './protocol/tracker_layer/instruction_message/InstructionMessageSerializerV2'
+import './protocol/tracker_layer/instruction_message/InstructionMessageSerializerV1'
 import PublishRequest from './protocol/control_layer/publish_request/PublishRequest'
 import './protocol/control_layer/publish_request/PublishRequestSerializerV1'
 import './protocol/control_layer/publish_request/PublishRequestSerializerV2'
@@ -27,13 +27,13 @@ import './protocol/control_layer/resend_response/ResendResponseResendingSerializ
 import './protocol/control_layer/resend_response/ResendResponseResendingSerializerV2'
 import ResendResponseResent from './protocol/control_layer/resend_response/ResendResponseResent'
 import StatusMessage from './protocol/tracker_layer/status_message/StatusMessage'
-import './protocol/tracker_layer/status_message/StatusMessageSerializerV2'
+import './protocol/tracker_layer/status_message/StatusMessageSerializerV1'
 import './protocol/control_layer/resend_response/ResendResponseResentSerializerV1'
 import './protocol/control_layer/resend_response/ResendResponseResentSerializerV2'
 import StorageNodesRequest from './protocol/tracker_layer/storage_nodes_request/StorageNodesRequest'
-import './protocol/tracker_layer/storage_nodes_request/StorageNodesRequestSerializerV2'
+import './protocol/tracker_layer/storage_nodes_request/StorageNodesRequestSerializerV1'
 import StorageNodesResponse from './protocol/tracker_layer/storage_nodes_response/StorageNodesResponse'
-import './protocol/tracker_layer/storage_nodes_response/StorageNodesResponseSerializerV2'
+import './protocol/tracker_layer/storage_nodes_response/StorageNodesResponseSerializerV1'
 import SubscribeRequest from './protocol/control_layer/subscribe_request/SubscribeRequest'
 import './protocol/control_layer/subscribe_request/SubscribeRequestSerializerV1'
 import './protocol/control_layer/subscribe_request/SubscribeRequestSerializerV2'
