+++ test/unit/protocol/tracker_layer/InstructionMessageSerializerV1.test.js
@@ -3,19 +3,30 @@ import assert from 'assert'
 import { TrackerLayer } from '../../../../src'
 import TrackerMessage from '../../../../src/protocol/tracker_layer/TrackerMessage'
 
-const { StatusMessage } = TrackerLayer
+const { InstructionMessage } = TrackerLayer
 
 const VERSION = 1
 
 // Message definitions
-const message = new StatusMessage({
+const message = new InstructionMessage({
     version: VERSION,
     requestId: 'requestId',
-    status: {}
+    streamId: 'streamId',
+    streamPartition: 10,
+    nodeAddresses: ['ws://address-1', 'ws://address-2'],
+    counter: 100
 })
-const serializedMessage = JSON.stringify([VERSION, TrackerMessage.TYPES.StatusMessage, 'requestId', {}])
+const serializedMessage = JSON.stringify([
+    VERSION,
+    TrackerMessage.TYPES.InstructionMessage,
+    'requestId',
+    'streamId',
+    10,
+    ['ws://address-1', 'ws://address-2'],
+    100
+])
 
-describe('StatusMessageSerializerV2', () => {
+describe('InstructionMessageSerializerV1', () => {
     describe('deserialize', () => {
         it('correctly parses messages', () => {
             assert.deepStrictEqual(TrackerMessage.deserialize(serializedMessage), message)
